+++ libs/angular/src/v-angular/modal/slide-out/slide-out.stories.ts
@@ -10,6 +10,8 @@ import {
 } from '@storybook/angular'
 
 import { NggCoreWrapperModule } from '@sebgroup/green-angular/src/lib/shared'
+import { DropdownUtils } from '../../core'
+import { NggvDropdownModule } from '../../dropdown/dropdown.module'
 import { NggvI18nModule } from '../../i18n/i18n.module'
 import { NggvSlideOutComponent } from './slide-out.component'
 
@@ -23,7 +25,7 @@ export default {
   component: NggvSlideOutComponent,
   decorators: [
     applicationConfig({
-      providers: [importProvidersFrom(NggvI18nModule)],
+      providers: [importProvidersFrom(NggvI18nModule), DropdownUtils],
     }),
     moduleMetadata({
       imports: [
@@ -31,14 +33,15 @@ export default {
         BrowserAnimationsModule,
         NggvI18nModule,
         NggCoreWrapperModule,
+        NggvDropdownModule,
       ],
     }),
   ],
   argTypes: {
     side: {
+      options: ['left', 'right'],
       control: {
         type: 'inline-radio',
-        options: ['left', 'right'],
       },
     },
     action: {
@@ -54,13 +57,74 @@ export default {
   },
 } as Meta
 
+const options = [
+  { key: 'opt1', label: 'label.defaultlabel', accountNumber: '345345' },
+  { key: 'opt6', label: 'HÖR LUR TJO HEJ KOTLETT' },
+  {
+    key: 'opt2',
+    label:
+      'This is a very long label that should over flow a smaller dropdown button',
+  },
+  { key: 'opt3', label: 'Option 3', disabled: true },
+  { key: 'opt4', label: 'Option 4' },
+  {
+    key: 'opt5',
+    label: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do',
+  },
+  {
+    label: 'Group',
+    options: [
+      { key: 'group-opt1', label: null, accountNumber: '3453245' },
+      { key: 'group-opt2', label: 'Option 2' },
+      { key: 'group-opt3', label: 'Option 3' },
+      { key: 'group-opt4', label: 'Option 4' },
+    ],
+  },
+]
+const dropdownArgs = {
+  required: false,
+  invalid: false,
+  label: 'Field label',
+  placeholder: 'Placeholder text',
+  description: 'This. Is. A. Description.',
+  ngModel: 'opt2',
+  options,
+}
+
 const Template: StoryFn<NggvSlideOutComponent & WithExtras> = (args: any) => ({
   template: /* html */ `
-  <nggv-slideout-modal #slideOut [initiallyShown]="initiallyShown" [closable]="closable" [content]="content" [title]="title" [buttons]="buttons" [side]="side" [autoWidth]="autoWidth" (nggvCloseEvent)="action($event)">
-      <p *ngIf="!outsideContent && !autoWidth">Some content in the content div</p>
+    <nggv-slideout-modal #slideOut
+      [initiallyShown]="initiallyShown"
+      [closable]="closable"
+      [content]="content"
+      [title]="title"
+      [buttons]="buttons"
+      [side]="side"
+      [autoWidth]="autoWidth"
+      (nggvCloseEvent)="action($event)"
+      >
+      <p *ngIf="!outsideContent && !autoWidth">
+        Some content in the content div:
+        <nggv-dropdown [required]="required"
+          [invalid]="invalid"
+          [label]="label"
+          [placeholder]="placeholder"
+          [description]="description"
+          [ngModel]="ngModel"
+          [options]="options">
+        </nggv-dropdown>
+      </p>
       <div *ngIf="outsideContent" slot="outside-content" [ngStyle]="{ 'background': 'black', 'color': 'white', 'padding': '1rem' }">
         This is the slotted content. This can be used in the (rare) case of content that needs
         to take up space outside the well-padded modal content div.
+        <nggv-dropdown [required]="required"
+          [invalid]="invalid"
+          [label]="label"
+          [placeholder]="placeholder"
+          [description]="description"
+          [ngModel]="ngModel"
+          [options]="options">
+        </nggv-dropdown>
       </div>
       <div *ngIf="autoWidth" slot="outside-content" [ngStyle]="{ 'background': 'green', 'color': 'white', 'padding': '1rem', 'width': '50vw' }">
         This is a modal with extra wide content.
@@ -84,6 +148,7 @@ Primary.args = {
   side: 'right',
   title: 'Modal title',
   content: 'Some content in the content div.\n\nSupports line breaks',
+  ...dropdownArgs,
   action: ((closed: boolean) => console.log('is closed?', closed)) as any,
 }